==> 12-AEM_WCRs_2_Texture-byDepth.ipynb <==
import pandas as pd
import geopandas as gpd
from shapely.geometry import Point
import os

# Load the data
filtered_df = pd.read_csv("8-updated_all.csv")

# Drop rows where INTERVALSTART is NaN
filtered_df = filtered_df.dropna(subset=['AverageCoarseFraction'])

filtered_df.head()

# Define the output folder path
output_folder = r"C:\Users\betebari\Documents\C2VSim_Texture\Texture-Maps"

# Ensure the output folder exists
if not os.path.exists(output_folder):
    os.makedirs(output_folder)

# Define the depth intervals you want to use

depth_intervals = [(0, 25), (25, 50), (50, 75),
                (75, 100), (100, 125), (125, 150)
                ]  # You can extend this as needed

# Loop through each depth interval and filter the data accordingly
for depth_range in depth_intervals:
    start_depth, end_depth = depth_range

    # Filter data for the current depth range
    depth_slice_df = filtered_df[
        (filtered_df['INTERVAL_MIDPOINT'] >= start_depth) & 
        (filtered_df['INTERVAL_MIDPOINT'] < end_depth)
    ]

    # Ensure there's data for this depth range
    if depth_slice_df.empty:
        print(f"No data available for depth range {start_depth}-{end_depth} feet")
        continue

    # Create geometry from UTMX and UTMY
    geometry = [Point(xy) for xy in zip(depth_slice_df['UTMX'], depth_slice_df['UTMY'])]

    # Create a GeoDataFrame using the point geometries
    gdf = gpd.GeoDataFrame(depth_slice_df, geometry=geometry)

    # Construct the output file name using the depth range
    shapefile_name = f"point_output_{start_depth}_{end_depth}.shp"

    # Full path for the shapefile
    shapefile_path = os.path.join(output_folder, shapefile_name)

    # Save the point data as a shapefile
    gdf.to_file(shapefile_path)

    print(f"Shapefile created: {shapefile_path}")
